HFP-1637 Removes transition of drop-down since it is not working anyway
It makes the iFrame resizing incorrect, because the height is not fully
expanded when it runs. Removing the transition for max-height allows us
to not have a setTimeout for when resizing. Additionally the transition
will never be seen anyway, because the iFrame is not big enough to show
it before the transition ends.